requirements.txt
sigma-cli==1.0.6
pyyaml==6.0.2